+++ spec/dummy/client/entries/app-kitten.js
@@ -39,6 +39,7 @@ import { TextInputWithUnit } from 'kitten/components/form/text-input-with-unit'
 
 // Grid
 import { Grid, GridCol } from 'kitten/components/grid/grid'
+import { KarlSideLayout } from 'kitten/karl/layout/side-layout'
 
 // Icons
 import GrabberIcon from 'kitten/components/icons/grabber-icon'
@@ -299,6 +300,7 @@ ReactOnRails.register({
   KarlGridInGrid,
   KarlGridSmall,
   KarlGridWithOffset,
+  KarlSideLayout,
   KarlParagraph,
   KarlParagraphSecondary,
   KarlParagraphTertiary,