+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `SideLayout` component and fix layout error on Safari.
+
 ## [9.0.0] - 2017-03-22
 
 Breaking change: